Fix setup.py for editable installations.
Add support for the --inplace option to setup.py build_ext. When this is
passed, compiled extensions are copied to the source directory. This is
where editable installations expect to find them, and pip install -e .
will pass the --inplace option to build_ext.

Fix the extension name (should be a package name, see
https://docs.python.org/3.7/distutils/apiref.html#distutils.core.Extension).

Fix a bug where an editable install on a fresh repository does not
create the lib.system folder in /build.