SubD iterators for all vertices/edges/faces of a SubD/vertex/edge/face
SubD iterators for all vertices/edges/faces, in or around a SubD/vertex/edge/face. Based on a standardization of all ON_SubDXXXIterators in a templated adapter class.

Define random access into these iterators. NB: Random access works differently when the iterator is from a SubD, in which case it refers to finding the vertex/edge/face with the given Id in the SubD.

Clean up BND_* classes related to iterators.