Improvements to CMakeLists.txt for Py MSVC: warnings, parallel builds
Get rid of warnings (CMP0148) when using CMake >= 3.21, related to finding the Python libraries.

Use parallel builds and the multi-tool task scheduler, as well as incremental linking, for faster builds.

Use OpenMP when building OpenNURBS, we do not need to limit ourselves to the single-threade version of that lib.

Prepare to build rhino3dm, ON and Zlib with warnings level 4 and Treat warnings as errors. For now, just enable /W3 in rhino3dm.